feat(TAP-9227): add apiserver alarm config

+++ manager/tm/src/main/java/com/tapdata/tm/v2/api/monitor/main/dto/ServerOverviewDetail.java
@@ -18,7 +18,7 @@ public class ServerOverviewDetail extends ValueBase {
 
     private Double cpuUsage;
 
-    private Long memoryUsage;
+    private Double memoryUsage;
 
     private Long usagePingTime;
